+++ tests/lib/managers/collaboration-whitelists-test.js
@@ -17,7 +17,7 @@ var BoxClient = require('../../../lib/box-client');
 // ------------------------------------------------------------------------------
 // Helpers
 // ------------------------------------------------------------------------------
-var sandbox = sinon.sandbox.create(),
+var sandbox = sinon.createSandbox(),
 	boxClientFake = leche.fake(BoxClient.prototype),
 	CollaborationWhitelist,
 	collaborationWhitelist,