+++ tests/lib/util/errors-test.js
@@ -14,7 +14,7 @@ var assert = require('assert'),
 // ------------------------------------------------------------------------------
 // Helpers
 // ------------------------------------------------------------------------------
-var sandbox = sinon.sandbox.create(),
+var sandbox = sinon.createSandbox(),
 	errors,
 	MODULE_FILE_PATH = '../../../lib/util/errors';
 
@@ -25,40 +25,41 @@ var sandbox = sinon.sandbox.create(),
 
 describe('Errors', function() {
 
-	before(function() {
-		// Enable Mockery
-		mockery.enable({ useCleanCache: true });
-		// Register Mocks
-		mockery.registerAllowable('http-status');
-		mockery.registerAllowable('util');
-		// Register Module Under Test
-		mockery.registerAllowable(MODULE_FILE_PATH);
-	});
-
 	beforeEach(function() {
-		// Setup File Under Test
+		mockery.enable({
+			warnOnUnregistered: false
+		});
+		mockery.registerAllowable(MODULE_FILE_PATH, true);
 		errors = require(MODULE_FILE_PATH);
 	});
 
 	afterEach(function() {
 		sandbox.verifyAndRestore();
-		mockery.resetCache();
-	});
-
-	after(function() {
 		mockery.deregisterAll();
 		mockery.disable();
 	});
 
-	describe('buildExpiredAuthError()', function() {
+	describe('buildAuthError()', function() {
 		it('should build an error object with the proper status code and message', function() {
 			var response = {
 				statusCode: 401
 			};
-			var errObject = errors.buildExpiredAuthError(response);
+			var errObject = errors.buildAuthError(response);
 			assert(errObject.authExpired);
 			assert.strictEqual(errObject.response, response);
-			assert.strictEqual(errObject.message, 'Expired Auth: Auth code or refresh token has expired.');
+			assert.strictEqual(errObject.message, 'Expired Auth: Auth code or refresh token has expired [401 Unauthorized]');
+		});
+
+		it('should use provided message when message argument is passed', function() {
+
+			var response = {
+				statusCode: 401
+			};
+			var message = 'test';
+			var errObject = errors.buildAuthError(response, message);
+			assert(errObject.authExpired);
+			assert.strictEqual(errObject.response, response);
+			assert.strictEqual(errObject.message, 'test [401 Unauthorized]');
 		});
 	});
 
@@ -70,7 +71,7 @@ describe('Errors', function() {
 			};
 			var errObject = errors.buildResponseError(response, 'testMessage');
 
-			assert.strictEqual(errObject.message, 'testMessage');
+			assert.strictEqual(errObject.message, 'testMessage [505 HTTP Version not Supported]');
 			assert.strictEqual(errObject.statusCode, 505);
 			assert.strictEqual(errObject.response, response);
 		});
@@ -99,7 +100,7 @@ describe('Errors', function() {
 			};
 
 			var errObject = errors.buildUnexpectedResponseError(response);
-			assert.strictEqual(errObject.message, 'Unexpected API Response [505 HTTP Version not Supported] (error_code: "Bad things happened")');
+			assert.strictEqual(errObject.message, 'Unexpected API Response [505 HTTP Version not Supported] error_code - Bad things happened');
 			assert.strictEqual(errObject.statusCode, 505);
 			assert.strictEqual(errObject.response, response);
 		});